deps/v8/test/webkit/fast/regex/pcre-test-4-expected.txt

# Copyright 2013 the V8 project authors. All rights reserved.
# Copyright (C) 2005, 2006, 2007, 2008, 2009 Apple Inc. All rights reserved.
#
# Redistribution and use in source and binary forms, with or without
# modification, are permitted provided that the following conditions
# are met:
# 1.  Redistributions of source code must retain the above copyright
#     notice, this list of conditions and the following disclaimer.
# 2.  Redistributions in binary form must reproduce the above copyright
#     notice, this list of conditions and the following disclaimer in the
#     documentation and/or other materials provided with the distribution.
#
# THIS SOFTWARE IS PROVIDED BY APPLE INC. AND ITS CONTRIBUTORS ``AS IS'' AND ANY
# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
# DISCLAIMED. IN NO EVENT SHALL APPLE INC. OR ITS CONTRIBUTORS BE LIABLE FOR ANY
# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
# ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

A chunk of our port of PCRE's test suite, adapted to be more applicable to JavaScript.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS regex0.exec(input0); is results
PASS regex0.exec(input1); is results
PASS regex0.exec(input2); is results
PASS regex0.exec(input3); is results
PASS regex1.exec(input0); is results
PASS regex1.exec(input1); is results
PASS regex1.exec(input2); is results
PASS regex1.exec(input3); is results
PASS regex1.exec(input4); is results
PASS regex2.exec(input0); is results
PASS regex3.exec(input0); is results
PASS regex4.exec(input0); is results
PASS regex5.exec(input0); is results
PASS regex6.exec(input0); is results
PASS regex7.exec(input0); is results
PASS regex8.exec(input0); is results
PASS regex9.exec(input0); is results
PASS regex10.exec(input0); is results
PASS regex11.exec(input0); is results
PASS regex12.exec(input0); is results
PASS regex12.exec(input1); is results
PASS regex12.exec(input2); is results
PASS regex12.exec(input3); is results
PASS regex12.exec(input4); is results
PASS regex13.exec(input0); is results
PASS regex13.exec(input1); is results
PASS regex13.exec(input2); is results
PASS regex13.exec(input3); is results
PASS regex13.exec(input4); is results
PASS regex13.exec(input5); is results
PASS regex14.exec(input0); is results
PASS regex14.exec(input1); is results
PASS regex14.exec(input2); is results
PASS regex14.exec(input3); is results
PASS regex14.exec(input4); is results
PASS regex14.exec(input5); is results
PASS regex15.exec(input0); is results
PASS regex15.exec(input1); is results
PASS regex15.exec(input2); is results
PASS regex15.exec(input3); is results
PASS regex15.exec(input4); is results
PASS regex15.exec(input5); is results
PASS regex15.exec(input6); is results
PASS regex15.exec(input7); is results
PASS regex15.exec(input8); is results
PASS regex16.exec(input0); is results
PASS regex16.exec(input1); is results
PASS regex16.exec(input2); is results
PASS regex16.exec(input3); is results
PASS regex16.exec(input4); is results
PASS regex16.exec(input5); is results
PASS regex16.exec(input6); is results
PASS regex16.exec(input7); is results
PASS regex16.exec(input8); is results
PASS regex17.exec(input0); is results
PASS regex21.exec(input0); is results
PASS regex21.exec(input1); is results
PASS regex22.exec(input0); is results
PASS regex22.exec(input1); is results
PASS regex22.exec(input2); is results
PASS regex30.exec(input0); is results
PASS regex31.exec(input0); is results
PASS regex32.exec(input0); is results
PASS regex33.exec(input0); is results
PASS regex34.exec(input0); is results
PASS regex35.exec(input0); is results
PASS regex36.exec(input0); is results
PASS regex37.exec(input0); is results
PASS regex37.exec(input1); is results
PASS regex38.exec(input0); is results
PASS regex38.exec(input1); is results
PASS regex39.exec(input0); is results
PASS regex39.exec(input1); is results
PASS regex40.exec(input0); is results
PASS regex41.exec(input0); is results
PASS regex42.exec(input0); is results
PASS regex43.exec(input0); is results
PASS regex44.exec(input0); is results
PASS regex45.exec(input0); is results
PASS regex47.exec(input0); is results
PASS regex48.exec(input0); is results
PASS regex49.exec(input0); is results
PASS regex49.exec(input1); is results
PASS regex50.exec(input0); is results
PASS regex50.exec(input1); is results
PASS regex51.exec(input0); is results
PASS regex52.exec(input0); is results
PASS regex53.exec(input0); is results
PASS regex53.exec(input1); is results
PASS regex54.exec(input0); is results
PASS regex54.exec(input1); is results
PASS regex54.exec(input2); is results
PASS regex54.exec(input3); is results
PASS regex55.exec(input0); is results
PASS regex55.exec(input1); is results
PASS regex55.exec(input2); is results
PASS regex55.exec(input3); is results
PASS regex56.exec(input0); is results
PASS regex57.exec(input0); is results
PASS regex57.exec(input1); is results
PASS regex57.exec(input2); is results
PASS regex58.exec(input0); is results
PASS regex58.exec(input1); is results
PASS regex58.exec(input2); is results
PASS regex59.exec(input0); is results
PASS regex59.exec(input1); is results
PASS regex60.exec(input0); is results
PASS regex60.exec(input1); is results
PASS regex60.exec(input2); is results
PASS regex61.exec(input0); is results
PASS regex61.exec(input1); is results
PASS regex61.exec(input2); is results
PASS regex62.exec(input0); is results
PASS regex63.exec(input0); is results
PASS regex63.exec(input1); is results
PASS regex63.exec(input2); is results
PASS regex64.exec(input0); is results
PASS regex64.exec(input1); is results
PASS regex64.exec(input2); is results
PASS regex65.exec(input0); is results
PASS regex65.exec(input1); is results
PASS regex66.exec(input0); is results
PASS regex66.exec(input1); is results
PASS regex66.exec(input2); is results
PASS regex67.exec(input0); is results
PASS regex67.exec(input1); is results
PASS regex67.exec(input2); is results
PASS regex68.exec(input0); is results
PASS regex68.exec(input1); is results
PASS regex69.exec(input0); is results
PASS regex69.exec(input1); is results
PASS regex69.exec(input2); is results
PASS regex70.exec(input0); is results
PASS regex70.exec(input1); is results
PASS regex70.exec(input2); is results
PASS regex71.exec(input0); is results
PASS regex71.exec(input1); is results
PASS regex71.exec(input2); is results
PASS regex72.exec(input0); is results
PASS regex72.exec(input1); is results
PASS regex72.exec(input2); is results
PASS regex73.exec(input0); is results
PASS regex73.exec(input1); is results
PASS regex74.exec(input0); is results
PASS regex74.exec(input1); is results
PASS regex74.exec(input2); is results
PASS regex75.exec(input0); is results
PASS regex75.exec(input1); is results
PASS regex75.exec(input2); is results
PASS regex75.exec(input3); is results
PASS regex75.exec(input4); is results
PASS regex76.exec(input0); is results
PASS regex76.exec(input1); is results
PASS regex76.exec(input2); is results
PASS regex77.exec(input0); is results
PASS regex77.exec(input1); is results
PASS regex77.exec(input2); is results
PASS regex77.exec(input3); is results
PASS regex78.exec(input0); is results
PASS regex78.exec(input1); is results
PASS regex78.exec(input2); is results
PASS regex78.exec(input3); is results
PASS regex78.exec(input4); is results
PASS regex78.exec(input5); is results
PASS regex79.exec(input0); is results
PASS regex79.exec(input1); is results
PASS regex79.exec(input2); is results
PASS regex80.exec(input0); is results
PASS regex80.exec(input1); is results
PASS regex80.exec(input2); is results
PASS regex81.exec(input0); is results
PASS regex81.exec(input1); is results
PASS regex81.exec(input2); is results
PASS regex82.exec(input0); is results
PASS regex82.exec(input1); is results
PASS regex82.exec(input2); is results
PASS regex86.exec(input0); is results
PASS regex86.exec(input1); is results
PASS regex86.exec(input2); is results
PASS regex86.exec(input3); is results
PASS regex86.exec(input4); is results
PASS regex87.exec(input0); is results
PASS regex87.exec(input1); is results
PASS regex87.exec(input2); is results
PASS regex87.exec(input3); is results
PASS regex87.exec(input4); is results
PASS regex87.exec(input5); is results
PASS regex88.exec(input0); is results
PASS regex88.exec(input1); is results
PASS regex88.exec(input2); is results
PASS regex88.exec(input3); is results
PASS regex88.exec(input4); is results
PASS regex91.exec(input0); is results
PASS regex91.exec(input1); is results
PASS regex91.exec(input2); is results
PASS regex91.exec(input3); is results
PASS regex91.exec(input4); is results
PASS regex92.exec(input0); is results
PASS regex93.exec(input0); is results
PASS regex94.exec(input0); is results
PASS regex95.exec(input0); is results
PASS regex95.exec(input1); is results
PASS regex96.exec(input0); is results
PASS regex97.exec(input0); is results
PASS regex98.exec(input0); is results
PASS regex99.exec(input0); is results
PASS regex100.exec(input0); is results
PASS regex102.exec(input0); is results
PASS regex103.exec(input0); is results
PASS regex104.exec(input0); is results
PASS regex105.exec(input0); is results
PASS regex105.exec(input1); is results
PASS regex106.exec(input0); is results
PASS regex107.exec(input0); is results
PASS input0.match(regexGlobal0); is results
PASS input1.match(regexGlobal0); is results
PASS input0.match(regexGlobal1); is results
PASS input0.match(regexGlobal2); is results
PASS input0.match(regexGlobal3); is results
PASS input0.match(regexGlobal4); is results
PASS input0.match(regexGlobal5); is results
PASS input0.match(regexGlobal6); is results
PASS input0.match(regexGlobal7); is results
PASS input0.match(regexGlobal8); is results
PASS input0.match(regexGlobal9); is results
PASS input0.match(regexGlobal10); is results
PASS input0.match(regexGlobal11); is results
PASS input0.match(regexGlobal12); is results
PASS input0.match(regexGlobal13); is results
PASS input0.match(regexGlobal14); is results
PASS input0.match(regexGlobal15); is results
PASS input0.match(regexGlobal16); is results
PASS input0.match(regexGlobal17); is results
PASS input0.match(regexGlobal19); is results
PASS input0.match(regexGlobal21); is results
PASS input0.match(regexGlobal24); is results
PASS input0.match(regexGlobal25); is results
PASS input0.match(regexGlobal26); is results
PASS input0.match(regexGlobal27); is results
PASS input0.match(regexGlobal28); is results
PASS input1.match(regexGlobal28); is results
PASS input0.match(regexGlobal29); is results
PASS input1.match(regexGlobal29); is results
PASS successfullyParsed is true

TEST COMPLETE
